Refactor: base class for batch_l2
