--- backpack/extensions/firstorder/batch_l2_grad/conv2d.py
@@ -1,6 +1,0 @@
-from backpack.extensions.firstorder.batch_l2_grad.convnd import BatchL2ConvND
-
-
-class BatchL2Conv2d(BatchL2ConvND):
-    def __init__(self):
-        super().__init__(N=2, params=["bias", "weight"])
