+++ fully_documented.txt
@@ -21,8 +21,7 @@ backpack/extensions/firstorder/variance/__init__.py
 backpack/extensions/firstorder/sum_grad_squared/sgs_base.py
 backpack/extensions/firstorder/sum_grad_squared/rnn.py
 backpack/extensions/firstorder/sum_grad_squared/__init__.py
-backpack/extensions/firstorder/batch_l2_grad/rnn.py
-backpack/extensions/firstorder/batch_l2_grad/__init__.py
+backpack/extensions/firstorder/batch_l2_grad/
 
 backpack/extensions/secondorder/diag_ggn/__init__.py
 backpack/extensions/secondorder/diag_ggn/diag_ggn_base.py
